backend/requirements.txt
#python3.10.9

fastapi
uvicorn[standard]
motor
python-dotenv
yfinance
pydantic[email]
passlib==1.7.4
bcrypt==3.2.0